refactor(conjugation): remove POC recipe APIs
