+++ src/polyzymd/builders/conjugation/polymer/recipe.py
@@ -8,16 +8,6 @@ from typing import Any
 
 from pydantic import AliasChoices, BaseModel, ConfigDict, Field, field_validator, model_validator
 
-POC_SBMA_SMILES = (
-    "[H]C([H])=C(C(=O)OC([H])([H])C([H])([H])[N+]"
-    "(C([H])([H])[H])(C([H])([H])[H])C([H])([H])C([H])([H])"
-    "C([H])([H])S(=O)(=O)[O-])C([H])([H])[H]"
-)
-POC_EGPMA_SMILES = (
-    "[H]C([H])=C(C(=O)OC([H])([H])C([H])([H])Oc1c([H])c([H])c([H])c([H])c1[H])C([H])([H])[H]"
-)
-POC_NHS_SMILES = "CC(=C)C(=O)ON1C(=O)CCC1=O"
-
 DEFAULT_PROBABILITY_TOLERANCE = 1.0e-6
 
 
@@ -238,7 +228,7 @@ class PolymerRecipe(BaseModel):
         return {monomer.name: monomer.residue_name for monomer in self.monomers}
 
 
-class PolymeristGenerationSmokeResult(BaseModel):
+class PolymeristGenerationResult(BaseModel):
     """Summary from the optional Polymerist recipe-generation boundary."""
 
     model_config = ConfigDict(arbitrary_types_allowed=True)
@@ -255,90 +245,14 @@ class PolymeristGenerationSmokeResult(BaseModel):
     rdkit_mol: Any | None = Field(default=None, exclude=True)
 
 
-def sbma_egpma_nhs_recipe(
-    *,
-    length: int = 10,
-    seed: int | None = 42,
-    reactive_monomer_index: int | None = None,
-    fixed_sequence: str | None = None,
-) -> PolymerRecipe:
-    """Build the SBMA/EGPMA/NHS recipe used by the conjugation POC.
-
-    Parameters
-    ----------
-    length : int, optional
-        Degree of polymerization, by default 10.
-    seed : int or None, optional
-        Random seed for deterministic sequence generation, by default 42.
-    reactive_monomer_index : int or None, optional
-        Zero-based NHS residue index. ``None`` centers NHS, by default ``None``.
-    fixed_sequence : str or None, optional
-        Exact monomer-label sequence overriding stochastic generation, by default
-        ``None``.
-
-    Returns
-    -------
-    PolymerRecipe
-        Validated stochastic polymer recipe.
-    """
-    return PolymerRecipe(
-        name="SBMA-EGPMA-NHS",
-        monomers=(
-            PolymerMonomerRecipe(
-                label="A",
-                name="SBMA",
-                residue_name="SBM",
-                smiles=POC_SBMA_SMILES,
-                probability=0.945,
-            ),
-            PolymerMonomerRecipe(
-                label="B",
-                name="EGPMA",
-                residue_name="EGP",
-                smiles=POC_EGPMA_SMILES,
-                probability=0.045,
-            ),
-            PolymerMonomerRecipe(
-                label="C",
-                name="NHS",
-                residue_name="NHS",
-                smiles=POC_NHS_SMILES,
-                probability=0.01,
-            ),
-        ),
-        length=length,
-        seed=seed,
-        reactive_monomer_label="C",
-        reactive_monomer_index=reactive_monomer_index,
-        fixed_sequence=fixed_sequence,
-    )
-
-
-def sbma_nhs_egpma_acb_recipe() -> PolymerRecipe:
-    """Build the deterministic v1 SBMA:NHS:EGPMA recipe.
-
-    Returns
-    -------
-    PolymerRecipe
-        Three-monomer recipe whose fixed sequence ``ACB`` maps to
-        SBMA:NHS:EGPMA with the NHS monomer centered for Lys linkage.
-    """
-    return sbma_egpma_nhs_recipe(
-        length=3,
-        seed=None,
-        reactive_monomer_index=1,
-        fixed_sequence="ACB",
-    )
-
-
-def generate_polymerist_smoke_polymer(
+def generate_polymerist_recipe_polymer(
     recipe: PolymerRecipe,
     cache_directory: Path | str,
     *,
     force_regenerate: bool = False,
     max_retries: int = 3,
     energy_minimize: bool = True,
-) -> PolymeristGenerationSmokeResult:
+) -> PolymeristGenerationResult:
     """Generate a small Polymerist-backed polymer artifact from a recipe.
 
     This boundary intentionally stops at Polymerist fragment and PDB generation.
@@ -361,7 +275,7 @@ def generate_polymerist_smoke_polymer(
 
     Returns
     -------
-    PolymeristGenerationSmokeResult
+    PolymeristGenerationResult
         Summary with generated sequence, cache paths, and object metadata.
     """
     from polyzymd.builders.fragment_generator import FragmentGenerator
@@ -425,7 +339,7 @@ def generate_polymerist_smoke_polymer(
     if atom_count is None:
         atom_count = _get_polymerist_atom_count(polymer_object)
 
-    return PolymeristGenerationSmokeResult(
+    return PolymeristGenerationResult(
         recipe_name=recipe.name,
         sequence=sequence,
         cache_directory=cache_path,
