+++ tests/test_conjugation_integrated_smoke.py
@@ -48,15 +48,13 @@ from polyzymd.builders.conjugation.pablo.parameterization import (
     create_interchange_from_pablo_topology,
 )
 from polyzymd.builders.conjugation.polymer.polymerist import generated_fragment_from_polymerist_pdb
-from polyzymd.builders.conjugation.polymer.recipe import (
-    generate_polymerist_smoke_polymer,
-    sbma_egpma_nhs_recipe,
-)
+from polyzymd.builders.conjugation.polymer.recipe import generate_polymerist_recipe_polymer
 from polyzymd.builders.conjugation.structure.pdb import (
     CrosslinkedPdbAssemblyOptions,
     PdbAtomRecord,
     write_crosslinked_pdb,
 )
+from tests._support.conjugation_polymer_recipes import sbma_egpma_nhs_recipe
 
 T = TypeVar("T")
 
@@ -94,7 +92,7 @@ def test_opt_in_integrated_conjugation_physics_smoke(tmp_path: Path):
     generation = _run_stage(
         "Polymerist SBMA/EGPMA/NHS generation",
         artifact_dir,
-        lambda: generate_polymerist_smoke_polymer(
+        lambda: generate_polymerist_recipe_polymer(
             recipe,
             artifact_dir / "polymerist-cache",
             force_regenerate=True,
